+++ structurizr-dsl/src/main/java/com/structurizr/dsl/FindRelationshipsParser.java
@@ -1,13 +1,12 @@
 package com.structurizr.dsl;
 
-import com.structurizr.model.Element;
 import com.structurizr.model.ModelItem;
 import com.structurizr.model.Relationship;
 
 import java.util.Set;
 import java.util.stream.Collectors;
 
-final class RelationshipsParser extends AbstractParser {
+final class FindRelationshipsParser extends AbstractParser {
 
     private static final String GRAMMAR = "!relationships <expression>";
 
@@ -24,7 +23,6 @@ final class RelationshipsParser extends AbstractParser {
         Set<ModelItem> modelItems = new ExpressionParser().parseExpression(expression, context);
 
         Set<Relationship> relationships = modelItems.stream().filter(mi -> mi instanceof Relationship).map(mi -> (Relationship)mi).collect(Collectors.toSet());
-
 
         if (relationships.isEmpty()) {
             throw new RuntimeException("No relationships found for expression \"" + expression + "\"");